+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## $v
+## v160702.65000
 
 - Adding `UniqueId` class and `uniqueId()` Facade.
 - Adding local `StrPad{}` test file.